[Bugfix] Preserve FP32 product precision in causal convolution
The prefill and update Triton kernels multiply BF16/FP16 operands before
adding to their FP32 accumulators. This rounds away product bits before
accumulation and can turn a representable cancellation residual into zero.
Widen both operands to FP32 at each multiply in both kernels, keeping the
existing output dtype, activation and convolution-history update behavior.

Add an exact cancellation regression through both public convolution APIs.
Cover BF16, FP16 and FP32, widths 2/3/4, a masked channel tail, and the
selected and null history slots. The expected residual is derived from
(1 + eps)^2 - (1 + 2*eps) = eps^2, with exactly representable test values.

Validation on AMD Radeon AI PRO R9700, ROCm 7.14, PyTorch 2.12.0 and
Triton 3.7.1, importing these upstream source files into the pinned vLLM
0.28 runtime: the original kernel fails all 12 BF16/FP16 regressions and
passes the six FP32 controls; the corrected file passes all 182 tests.
All applicable pre-commit checks pass. CUDA and XPU were not tested here.

+++ vllm/model_executor/layers/mamba/ops/causal_conv1d.py
@@ -452,7 +452,8 @@ def _causal_conv1d_fwd_kernel(  # continuous batching
                     x_ptrs_1d = x_base_1d + idx_token * stride_x_token  # [BLOCK_N]
                     matrix_x = tl.load(x_ptrs_1d, mask=mask_x_1d)
 
-            acc += matrix_x * matrix_w  # [BLOCK_N]
+            # Widen before multiplying to avoid rounding each product to x.dtype.
+            acc += matrix_x.to(tl.float32) * matrix_w.to(tl.float32)  # [BLOCK_N]
 
         if KERNEL_WIDTH == 2:
             col0 = matrix_x
@@ -1058,7 +1059,8 @@ def _causal_conv1d_update_kernel(
                     x_ptrs_1d = x_base_1d + idx_token * stride_x_token  # [BLOCK_N]
                     matrix_x = tl.load(x_ptrs_1d, mask=mask_x_1d)
 
-            acc += matrix_x * matrix_w  # [BLOCK_N]
+            # Widen before multiplying to avoid rounding each product to x.dtype.
+            acc += matrix_x.to(tl.float32) * matrix_w.to(tl.float32)  # [BLOCK_N]
 
         if KERNEL_WIDTH == 2:
             col0 = matrix_x
